feat(homefeed): Add a RecyclerView component to the home_fragment
This will house all the posts

refactor(homefeed): Remove the pre-existing code in HomeFragment.java

All the stuff with the ViewModel, which we won't need now. I think.

refactor(homefeed): Delete autogenerated textView in fragment_home.xml

feat(homefeed): Put code for RecyclerView in HomeFragment

This is not compilable, as the adapter hasn't yet been written