fix(bottomNav): Disable the bottom bar when in a comments or post detail view
This prevents the bug that Sali showed with the icons highlight not updating.

refactor(bottomNav): Use argument-based logic instead of hard coding ID's

Instead of hardcoding layout ID's, add an argument to the fragments which don't need the appbar. This approach is more flexible as you only need to change the navigation graph instead of the java code.